[ft transfer benchmark] Restart locust script (#11391)
This PR aims to MVP FT transfer benchmark (#11348). This `.sh` script
restarts locust and neard if there are changes on remote. It will be
runned on one of our VMs. Later it probably should be triggered by CI
job, but for now we can just make a cron job which will run it time to
time.

Co-authored-by: Viktar Makouski <viktar@neaar.org>