refactor: Simplify retrieval of transaction id (#11410)
The transaction id is available after its creation, which we want to use
in https://github.com/near/nearcore/pull/11364. It was tricky to extract
it because we were receiving a serialized transaction from this API, so
I refactored the code to return SignedTransaction instead which also now
contains transaction_id.