perf: Migrate to pynacl.SigningKey (#11409)
We use it extensively in Locust benchmarks and it shows as a hotspot. A
single signature takes around 1-2ms on my machine. PyNacl is a
recommended replacement for python-ed25519 that is supposed to be 10-20
times faster for signatures.

This is a part of https://github.com/near/nearcore/issues/11352

I've followed this guide:
https://github.com/warner/python-ed25519?tab=readme-ov-file#migrating-to-pynacl