feat(congestion_control) - congestion info validation  (#11408)
Added the validation of the congestion info fields. As far as I can tell
this should cover both stateful and stateless validation transition. I
added a unit test for the relevant method as well but I am not actually
convinced that this is sufficient. I will see if I can add a simple
nayduck end to end test as well. I also checked that the test will pass
after cc stabilization.