fix(forknet): dont set state_sync.dump.location to null (#11386)
if --gcs-state-sync is not provided, we set this config parameter to
null, but it leads to an error deserializing the config. so only set
non-null state sync locations. Also warn the user if this is happening,
because if there's a state dumper node configured but this option wasn't
given, probably you did something wrong