Panic in debug mode or non-prod chain when there is a chunk-witness validation error (#11412)
This is an attempt to create a submittable version of [this
commit](https://github.com/near/nearcore/commit/0c6c82266302c3b115d73c4df66cdc545ba37d37).

After this change chunk-witness validation errors will lead to panic for
non-production chains (eg. mocknet and localnet). mainnet and testnet
will continue to log error and continue on validation errors.
We also add a method to enable/disable this for tests, especially for
adversarial behavior where the chunk validation errors are intended as
part of the test scenario.