[Memtrie] Implement parallel loading algorithm for memtrie. (#11356)
This PR implements parallel loading of memtries, using a small number of
nodes in the State column to determine the split points. It massively
improves loading speed. When tested on a 16 core GCP machine, this
decreases memtrie loading time from 122s to 13s.

From the top of parallel_loader.rs:
```
/// Logic to load a memtrie in parallel. It consists of three stages:
///  - First, we use the State column to visit the trie starting from the root. We recursively
///    expand the trie until all the unexpanded subtrees are small enough
///    (memory_usage <= `subtree_size`). The trie we have expanded is represented as a "plan",
///    which is a structure similar to the trie itself.
///  - Then, we load each small subtree (the keys under which all share a common prefix) in
///    parallel, by reading the FlatState column for keys that correspond to the prefix of that
///    subtree. The result of each construction is a `MemTrieNodeId` representing the root of that
///    subtree.
///  - Finally, We construct the final trie by using the loaded subtree roots and converting the
///    plan into a complete memtrie, returning the final root.
```

This PR includes the parallel loader, a concurrent arena to support the
parallel loading step, and updates to use parallel loading for initial
load only. As a bonus, the previous code that supports deferred
computation of node hash (which is complex and contains an unsafe block)
is deleted, since it is no longer necessary to defer it (the original
purpose was to parallelize the hash computation, but now the whole thing
is parallelized).