feat(debug): add congestion control debug page (#11442)
Adding a new debug page for congestion control at
`/debug/pages/congestion_control`.
It shows the list of most recent blocks and chunks. There is information
about various parameters that can tell the level of congestion across
shards.

No data is shown if the congestion control feature is not enabled.