feat(congestion): reject new transactions on RPC level (#11419)
Summary: In this PR, we introduce a new failure mode on the RPC level
when a transaction is submitted under congestion. The error is of type
`InvalidTxError` and called `ShardCongested` with a single field
`shard_id` referencing the congested shard.

## Details

With [cross-shard congestion
control](https://github.com/near/NEPs/pull/539) being stabilized soon,
we must deal with the case when a shard rejects new transactions.

On the chunk producer level, all transactions going to a congested shard
will be dropped. This keeps the memory requirements of chunk producers
bounded. Further, we decided to go for a relatively low threshold in
order to keep the latency of accepted transactions low, preventing new
transactions as soon as we hit 25% congestion on a specific shard.
Consequently, when shards are congested, it will not be long before
transactions are rejected.

This has consequences for the users. On the positive side, they will no
longer have to wait for a long time not knowing if their transaction
will be accepted or not. Either, it is executed within a bounded time
(at most 20 blocks after inclusion) or it will be rejected immediately.

But on the negative side, when a shard is congested, they will have to
actively retry sending the transaction until it gets accepted. We hope
that this can be automated by wallets, which can also provide useful
live updates to the user about what is happening. But for this, they
will need to understand and handle the new error `ShardCongested`
differently from existing errors. The key difference is that the same
signed transaction can be sent again and will be accepted if congestion
has gone down.