test(stateless_validation) - Implemented a protocol upgrade integration test (#11472)
This test performs a protocol upgrade from the version before
StatelessValidationV0 to the binary's PROTOCOL_VERSION. This test would
have caught the issue fixed in #11469. As is, in nightly, the test
checks protocol upgrade from 79 to 143.

Note that we also have protocol upgrade tests for congestion control. I
think this test didn't catch the mentioned issue because it's configured
with only one client and (I think) it automatically accepts its own
chunks. It was much easier to reuse the stateless validation test
configuration than to reimplement it in congestion control tests. Also
this test seems generally useful.