config(congestion_control) - set max_congestion_missed_chunks to 2 (#11473)
Setting the max_congestion_missed_chunks to 2. This means that at most 2
(+/-1) chunks worth of outgoing receipts can accumulate before all
shards will stop forwarding any receipts to the shard with missing
chunks. This is a relatively conservative limit to minimize the impact
of incoming receipts on source_receipt_proofs size in the state witness.
I expect it to have minimal impact on the total throughput of the chain
since it's quite rare to see more than one missed chunk in a row.

It's also worth noting that just having one missed chunk is a special
case that will not increase the congestion level:

https://github.com/near/nearcore/blob/dd1e2786bac3622020f6e684e986eb426751600a/core/primitives/src/congestion_info.rs#L65-L68