feat: epoch analysis tool (#11343)
Tool to analyse epoch infos in two modes:
* `check-consistency` - regenerate next next epoch info based on two
previous epochs and check that it matches the epoch info stored in DB;
* `backtest` - regenerate epoch infos with existing proposals, rewards
and kickouts as if `PROTOCOL_VERSION` was always in place.

The `backtest` was used to estimate new algorithm for chunk producer
shard assignments and showed that on average there is only one state
sync happening, if we use start epoch height >= 545. Epoch info for 544
can't be retrieved for some reason, see #11477.

Consistency check revealed that some epochs in the past can't be
replayed, see #11476.
Expected output of it, note that epoch T generates epoch T+2:
```
$ neard view-state epoch-analysis --start-height 1359 check-consistency
HEIGHT | VERSION | STATE SYNCS
  1361 |      53 |           0
  1362 |      53 |           8
  1363 |      54 |           8
  1364 |      54 |           8
  1365 |      54 |          24
  1366 |      54 |          12
...
```