Feat(eth-wallet-contract): automatic storage deposit during ERC-20 emulation (#11435)
This PR changes the wallet contract implementation in two ways:

(1) a gas optimization to emulated base-token transfers where the
address registrar contract needs to be consulted less often;
(2) a change to ERC-20 token emulation where transfers now check if the
receiver is registered with the NEP-141 contract that is acting as an
ERC-20.

The latter change is the main motivation for the PR because otherwise it
is otherwise inconvenient for users to use their Ethereum wallet
interface to transfer their tokens to other users with only Ethereum
wallets (without this PR those transactions would fail if the receiving
account had never held that kind of token before, but on Ethereum such a
transaction would succeed). However, the former change introduces some
code refactoring which actually made the second change easier to
implement, so it was a double win.

For testing (1) we rely on the existing tests for base token transfers
and relayers trying to incorrectly set the `target`. For (2) we modify
and extend the existing ERC-20 test to include both the case where the
receiver has not been registered with the NEP-141 and when it has.