refactor: split runtime::apply into smaller functions (#11490)
This is an attempt at refactoring `Runtime::apply`. The function was
originally very big, so the strategy here is to split it into smaller
chunks. The function signature is not touched.
I decided to add a few wrapper `structs` to help share the state which
before was all contained into `Runtime::apply`'s scope.

The PR contains multiple iterative commits hoping that it might help
reviewing.