feat: adjust witness size limits (#11511)
This PR includes the following:
* Set size limit for compressed witness to 32MiB. This is at least twice
as big as the witness we expected to handle, so it give us quite some
safety margin.
* Set size limit for the compressed witness to 64MiB. Same as above,
this values give us pretty big margin on top of witness sizes we expect.
* Reduce partial witness cache size to 40 entries to limit memory usage
at ~1.3GiB. 40 entries should be more than enough to handle all kinds of
crazy chain forks 🙃