fix(nayduck) - fix congestion control test - configure timeouts  (#11542)
The default timeouts introduced in #11397 mean that a failing requests
will take about 1 minute to complete including the timeouts. In the
congestion control test we intentionally overload the chain and expect
that some transactions will be rejected. Combined this means that the
test would take 20 minutes to complete. The fix is to better configure
the timeouts and retries in the nayduck infra.