Add a metric showing what limited number of processed receipts in a chunk (#11568)
The metric counts the number of chunks where the number of processed
receipts was limited by a certain factor.

Here's an example visualisation from a localnet node with ft.py loadtest
at ~500 TPS:


![image](https://github.com/near/nearcore/assets/149345204/b2cdcd51-e620-49e3-9507-6282a36144a7)

We can see that for ~20% of chunks on shard 0 the number of processed
receipts was limited by the `compute_limit`.

This will allow us to see what's the bottleneck for receipt processing -
is it the compute limit, storage proof limit, or something else.