runtime: implement loading the contract through Externals (#11550)
This simplifies the interface to the contract runtime and makes it
easier to split up the loading vs. instantiation vs. execution as it
gets exposed as an API from the contract runtime.

Thematically `VMContext` is probably more appropriate place for storing
some of this functionality, however Externals is quite a bit more of a
straightforward home for this functionality by the virtue of it already
having access to the storage where the contracts live in.

The downside of this being stored in `Externals` is precisely that until
this point `Externals` has been entirely focused on providing an
interface
for host functions to interact with. And now the trait is gaining
methods
that are slightly outside of that area of responsibility.