wasmtime: add caching of executables and artifacts (#11532)
This largely mirrors the code in near_vm_runner module. I heard some
people pondering what it would be like to use a higher quality backend.
Outside LLVM, Cranelift is by far the next in line in produced code
quality. Since we already have wasmtime in place, might as well wire it
up completely for a full experience.

Based on top of #11529
Part of #11319