fix: make TestnetFewerBlockProducers only apply to testnet (#11626)
Make `TestnetFewerBlockProducers` only apply to testnet to avoid
breaking other testing network due to the hardcoded change of block
producers. Specifically, increasing the number of block producers will
make things break due to `num_total_parts` computation that only looks
at what is in genesis and if genesis protocol version is between the
protocol version of `TestnetFewerBlockProducers` and
`NoChunkOnlyProducers`, it will forcefully set the number of block
producer seats to 20, regardless of what is set in genesis. Even worse,
any attempt to increase that later in the code won't work.