refactor: replace some references with owned structures in VMLogic (#11615)
`VMLogic` containing a lifetime makes it difficult to have it live for
any longer than a short sequence of `instantiate-link-run` and is one of
the reasons why we're forced to have some unsafe code in our linking
code.

This refactor replaces some of the reference fields with `Arc`s,
`Box`es, etc. This is not a complete refactor, I intend to do the
remainder as a follow-up.

Based on #11614
Part of #11319